@@ -1,21 +1,6 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
 import App from './App.vue'
-import VueRouter from 'vue-router'
-import routes from './routes'
-Vue.use(VueRouter)
+import router from './router'
+import store from './store'
 
-Vue.config.productionTip = false
-
-const router = new VueRouter({
-    mode: 'history',
-    routes
-})
-
-
-new Vue({
-    router, 
-  render: h => h(App),
-}).$mount('#app')
-
-
-
+createApp(App).use(store).use(router).mount('#app')
